requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
pdfplumber==0.10.4
PyMuPDF==1.23.21
pytesseract==0.3.10
Pillow==10.2.0
nltk==3.8.1
sentence-transformers==2.5.1
faiss-cpu==1.7.4
rank_bm25==0.2.2
google-genai==0.8.0
streamlit==1.31.1
numpy==1.26.4
pandas==2.2.0
pytest==8.0.0
pytest-asyncio==0.23.5
ragas==0.1.1
python-dotenv==1.0.1
httpx==0.26.0
